requirements.txt
fastapi[standard]
uvicorn[standard]
jinja2
motor
pydantic
python-dotenv
email-validator
slowapi
mailjet_rest
PyJWT